feat: filtering Component
