+++ src/main.jsx
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom/client'
 import Header from './components/main/Header'
 import './index.css'
 import Footer from './components/main/Footer'
-import Filtering from './components/main/Filtering'
+import Filtering from './components/main/Filtering/Filtering'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
